fix bug

+++ Src/CollisionPrimitive/Ray3d.h
@@ -37,7 +37,7 @@ public:
 
 	static bool		RayIntersectAABB_1D(float start, float dir, float min, float max, float* enter, float* exit)
 	{
-		if (fabs(dir) < 0.000001f)
+		if (fabsf(dir) < 0.000001f)
 		{
 			return start >= min && start <= max;
 		}
